Upload_project/requirements.txt
# altair==5.1.1
# appdirs==1.4.4
# argon2-cffi==21.3.0
# argon2-cffi-bindings==21.2.0
# asgiref==3.5.2
# asttokens==2.0.7
# attrs==22.1.0
# backcall==0.2.0
# beautifulsoup4==4.11.1
# bleach==5.0.1
# blinker==1.6.2
# cachetools==5.3.1
# certifi==2022.9.24
# cffi==1.15.1
# charset-normalizer==2.1.1
# click==8.1.7
# colorama==0.4.5
# cryptography==38.0.1
# cycler==0.11.0
# debugpy==1.6.2
# decorator==5.1.1
# defusedxml==0.7.1
# dill==0.3.5.1
# distro==1.8.0
Django==4.1.4
djangorestframework==3.14.0
# docplex==2.25.236
# entrypoints==0.4
# executing==0.9.1
# fastdtw==0.3.4
# fastjsonschema==2.16.1
# fonttools==4.36.0
# frozendict==2.3.8
# gitdb==4.0.10
GitPython==3.1.35
# html5lib==1.1
# idna==3.4
# importlib-metadata==6.8.0
# inflection==0.5.1
# ipykernel==6.15.1
# ipython==8.4.0
# ipython-genutils==0.2.0
# ipywidgets==7.7.1
# jedi==0.18.1
# Jinja2==3.1.2
# joblib==1.2.0
# JPype1==1.4.1
# jsonschema==4.9.1
# jupyter==1.0.0
# jupyter-client==7.3.4
# jupyter-console==6.4.4
# jupyter-core==4.11.1
# jupyterlab-pygments==0.2.2
# jupyterlab-widgets==1.1.1
# kiwisolver==1.4.4
# lxml==4.9.1
# markdown-it-py==3.0.0
# MarkupSafe==2.1.1
# matplotlib==3.5.3
# matplotlib-inline==0.1.3
# mdurl==0.1.2
# mistune==0.8.4
# more-itertools==10.1.0
# mpmath==1.2.1
# multitasking==0.0.11
# Nasdaq-Data-Link==1.0.4
# nbclient==0.6.6
# nbconvert==6.5.3
# nbformat==5.4.0
# nest-asyncio==1.5.5
# networkx==3.1
# notebook==6.4.12
# ntlm-auth==1.5.0
# numpy==1.23.2
# packaging==21.3
# pandas==1.4.3
# pandas-datareader==0.10.0
# pandocfilters==1.5.0
# parso==0.8.3
# pbr==5.10.0
# pickleshare==0.7.5
# Pillow==9.2.0
# ply==3.11
# prometheus-client==0.14.1
# prompt-toolkit==3.0.30
# protobuf==4.21.5
# psutil==5.9.1
# pure-eval==0.2.2
# pyarrow==13.0.0
# pycparser==2.21
# pydeck==0.8.0
# pyftpdlib==1.5.7
# Pygments==2.16.1
# Pympler==1.0.1
# Pyomo==6.6.1
# pyparsing==3.0.9
# PyPDF2==3.0.1
# pyrsistent==0.18.1
# python-dateutil==2.8.2
# python-pptx==0.6.22
# pytz==2023.3.post1
# pytz-deprecation-shim==0.1.0.post0
# pywin32==304
# pywinpty==2.0.7
# pyzmq==23.2.0
# qiskit==0.38.0
# qiskit-aer==0.11.0
# qiskit-finance==0.3.4
# qiskit-ibmq-provider==0.19.2
# qiskit-machine-learning==0.4.0
# qiskit-optimization==0.5.0
# qiskit-terra==0.25.1
# qtconsole==5.3.1
# QtPy==2.2.0
# requests==2.31.0
# requests-ntlm==1.1.0
# retworkx==0.11.0
# rich==13.5.2
# rustworkx==0.13.1
# scikit-learn==1.1.2
# scipy==1.9.0
# seaborn==0.11.2
# Send2Trash==1.8.0
# six==1.16.0
# smmap==5.0.0
# soupsieve==2.3.2.post1
# sqlparse==0.4.3
# stack-data==0.3.0
# stevedore==4.0.0
# streamlit==1.26.0
# sympy==1.11.1
# tabula-py==2.8.1
# tenacity==8.2.3
# tensorflow-hub==0.12.0
# terminado==0.15.0
# threadpoolctl==3.1.0
# tinycss2==1.1.1
# toml==0.10.2
# toolz==0.12.0
# tornado==6.2
# traitlets==5.3.0
# tweedledum==1.1.1
# typing_extensions==4.7.1
# tzdata==2022.7
# tzlocal==4.3.1
# urllib3==1.26.12
# validators==0.22.0
# watchdog==3.0.0
# wcwidth==0.2.5
# webencodings==0.5.1
# websocket-client==1.4.1
# websockets==10.3
# widgetsnbextension==3.6.1
# XlsxWriter==3.1.3
# yfinance==0.2.28
# zipp==3.16.2